Create TradingBotDeployment
The TradingBotDeployment Script is essentially a trading bot that executes and manages trades. It involved taking the predictions from a Neural Network in order to place trades through an API. The trading bot also managed any open positions and is also capable of closing trades if the requirements are met to do so.